create TorchLLM and TrtLLM
